Customer

+++ lib/services/order_service.dart
@@ -8,15 +8,22 @@ import 'package:picpee_mobile/services/auth_service.dart';
 
 class OrderService {
   // Lấy list orders của project
-  Future<List<OrderModel>> fetchOrders(int projectId) async {
+  Future<List<OrderModel>> fetchOrders({int? projectId}) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
+    Map<String, dynamic> body = {};
+    if (projectId != null) {
+      body['projectId'] = projectId;
+    }
     final response = await http.post(
       Uri.parse(Url.getOrdersOfProject),
       headers: {
         'Authorization': 'Bearer $token',
         'Content-Type': 'application/json',
       },
-      body: jsonEncode({'projectId': projectId}),
+      body: jsonEncode(body),
     );
     if (response.statusCode == 200) {
       final res = jsonDecode(response.body);
@@ -33,6 +40,9 @@ class OrderService {
   // Lấy chi tiết order
   Future<OrderModel?> fetchOrderDetails(int orderId) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.get(
       Uri.parse("${Url.getDetailOrder}/$orderId"),
       headers: {
@@ -54,6 +64,9 @@ class OrderService {
   // Tạo order mới
   Future<OrderModel> createOrder(Map<String, dynamic> orderData) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.post(
       Uri.parse(Url.createOrder),
       headers: {
@@ -71,9 +84,39 @@ class OrderService {
     }
   }
 
+  //Thanh toán order
+  Future<OrderModel> payOrder(int orderId, String? code) async {
+    final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
+    Map<String, dynamic> body = {};
+    if (code != null && code.isNotEmpty) {
+      body['code'] = code;
+    }
+    final response = await http.put(
+      Uri.parse("${Url.payOrder}/$orderId"),
+      headers: {
+        'Authorization': 'Bearer $token',
+        'Content-Type': 'application/json',
+      },
+      body: jsonEncode(body),
+    );
+    if (response.statusCode == 200) {
+      final res = jsonDecode(response.body);
+      final data = res['data'];
+      return OrderModel.fromJson(data);
+    } else {
+      throw Exception('Failed to pay order');
+    }
+  }
+
   // Hoàn thành order
   Future<bool> completeOrder(int orderId) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.put(
       Uri.parse("${Url.completedOrder}/$orderId"),
       headers: {
@@ -92,6 +135,9 @@ class OrderService {
   //Dispute order
   Future<bool> disputeOrder(int orderId, String reason) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.put(
       Uri.parse("${Url.disputeOrder}/$orderId"),
       headers: {
@@ -110,6 +156,9 @@ class OrderService {
   // Revision order
   Future<bool> revisionOrder(int orderId, String note) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.put(
       Uri.parse("${Url.revisionOrder}/$orderId"),
       headers: {
@@ -125,9 +174,33 @@ class OrderService {
     }
   }
 
+  //delete order
+  Future<bool> deleteOrder(int orderId) async {
+    final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
+    final response = await http.put(
+      Uri.parse("${Url.deleteOrder}/$orderId"),
+      headers: {
+        'Authorization': 'Bearer $token',
+        'Content-Type': 'application/json',
+      },
+      body: jsonEncode({}),
+    );
+    if (response.statusCode == 200) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+
   //Lấy order activity
   Future<List<OrderActivityModel>> fetchOrderActivities(int orderId) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.get(
       Uri.parse("${Url.getActivity}/$orderId"),
       headers: {
@@ -150,6 +223,9 @@ class OrderService {
   //Lấy comments
   Future<List<CommentModel>> fetchOrderComments(int orderId) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.post(
       Uri.parse(Url.getComment),
       headers: {
@@ -160,10 +236,12 @@ class OrderService {
     );
     if (response.statusCode == 200) {
       final res = jsonDecode(response.body);
-      final data = res['data'];
+      final data = res['data']['list'];
+      print(data.length);
       List<CommentModel> comments = (data as List)
           .map((commentJson) => CommentModel.fromJson(commentJson))
           .toList();
+      print(comments.length);
       return comments;
     } else {
       throw Exception('Failed to load order comments');
@@ -173,6 +251,9 @@ class OrderService {
   // Thêm comment
   Future<bool> addComment(Map<String, dynamic> commentData) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.post(
       Uri.parse(Url.addComment),
       headers: {
@@ -184,6 +265,7 @@ class OrderService {
     if (response.statusCode == 200) {
       return true;
     } else {
+      print("Failed to add comment: ${response.body}");
       return false;
     }
   }
@@ -191,6 +273,9 @@ class OrderService {
   // Xoá comment
   Future<bool> deleteComment(int commentId) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.delete(
       Uri.parse("${Url.addComment}/$commentId"),
       headers: {
@@ -208,6 +293,9 @@ class OrderService {
   // Cập nhật comment
   Future<bool> updateComment(int id, Map<String, dynamic> commentData) async {
     final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
     final response = await http.put(
       Uri.parse("${Url.addComment}/$id"),
       headers: {
@@ -219,7 +307,33 @@ class OrderService {
     if (response.statusCode == 200) {
       return true;
     } else {
+      print("Failed to update comment: ${response.body}");
       return false;
+    }
+  }
+
+  //lấy checklist
+  Future<List<OrderAddOn>> fetchChecklist(int orderId) async {
+    final token = await AuthService().getToken();
+    if (token == null) {
+      throw Exception('User not logged in');
+    }
+    final response = await http.get(
+      Uri.parse("${Url.getCheckList}/$orderId/get-check-list"),
+      headers: {
+        'Authorization': 'Bearer $token',
+        'Content-Type': 'application/json',
+      },
+    );
+    if (response.statusCode == 200) {
+      final res = jsonDecode(response.body);
+      final data = res['data'];
+      List<OrderAddOn> checklist = (data as List)
+          .map((checklistJson) => OrderAddOn.fromJson(checklistJson))
+          .toList();
+      return checklist;
+    } else {
+      throw Exception('Failed to load checklist');
     }
   }
 }
